\name{acvs.andel8}
\alias{acvs.andel8}
\title{Autocovariance and Autocorrelation Sequences for a Seasonal
  Persistent Process}
\description{
  The autocovariance and autocorrelation sequences from the time series
  model in Figure 8 of Andel (1986).  They were obtained through numeric
  integration of the spectral density function.
}
\usage{data(acvs.andel8)
}
\format{
  A data frame with 4096 rows and three columns:
  \item{lag}{Lag}
  \item{acvs}{Autocovariance sequence}
  \item{acf}{Autocorrelation sequence}
}
\references{
  Andel, J. (1986)
  Long memory time series models,
  \emph{Kypernetika}, \bold{22}, No. 2, 105-123.
}
\keyword{datasets}
