\name{Andel}
\alias{acvs.andel8}
\alias{acvs.andel9}
\alias{acvs.andel10}
\alias{acvs.andel11}
\title{Autocovariance and Autocorrelation Sequences for a Seasonal
  Persistent Process}
\description{
  The autocovariance and autocorrelation sequences from the time series
  model in Figures 8, 9, 10, and 11 of Andel (1986).  They were obtained
  through numeric integration of the spectral density function.
}
\usage{data(acvs.andel8)
data(acvs.andel9)
data(acvs.andel10)
data(acvs.andel11)
}
\format{
  A data frame with 4096 rows and three columns: lag, autocovariance sequence,
  autocorrelation sequence.
}
\references{
  Andel, J. (1986)
  Long memory time series models,
  \emph{Kypernetika}, \bold{22}, No. 2, 105-123.
}
\keyword{datasets}
