% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ar1}
\alias{ar1}
\title{Simulated AR(1) Series}
\format{
A vector containing 200 observations.
}
\usage{
data(ar1)
}
\description{
Simulated AR(1) series used in Gencay, Selcuk and Whitcher (2001).
}
\references{
Gencay, R., F. Selcuk and B. Whitcher (2001) \emph{An
Introduction to Wavelets and Other Filtering Methods in Finance and
Economics}, Academic Press.
}
\keyword{datasets}
