% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barbara}
\alias{barbara}
\title{Barbara Test Image}
\format{
A 256 \eqn{\times}{x} 256 matrix.
}
\source{
Internet.
}
\usage{
data(barbara)
}
\description{
The Barbara image comes from Allen Gersho's lab at the University of
California, Santa Barbara.
}
\keyword{datasets}
