% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blocks}
\alias{blocks}
\title{A Piecewise-Constant Function}
\format{
A vector containing 512 observations.
}
\source{
S+WAVELETS.
}
\usage{
data(blocks)
}
\description{
\deqn{blocks(x) = \sum_{j=1}^{11}(1 + {\rm sign}(x-p_j)) h_j / 2}{%
blocks(x) = sum[j=1,11] (1 + sign(x - p_j)) h_j/2}
}
\references{
Bruce, A., and H.-Y. Gao (1996) \emph{Applied Wavelet Analysis
with S-PLUS}, Springer: New York.
}
\keyword{datasets}
