% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doppler}
\alias{doppler}
\title{Sinusoid with Changing Amplitude and Frequency}
\format{
A vector containing 512 observations.
}
\source{
S+WAVELETS.
}
\usage{
data(doppler)
}
\description{
\deqn{doppler(x) = \sqrt{x(1 - x)} }{% doppler(x) = sqrt{x(1-x)}
sin[(2.1*pi)/(x+0.05)]}\deqn{ \sin\left(\frac{2.1\pi}{x+0.05}\right)}{%
doppler(x) = sqrt{x(1-x)} sin[(2.1*pi)/(x+0.05)]}
}
\references{
Bruce, A., and H.-Y. Gao (1996) \emph{Applied Wavelet Analysis
with S-PLUS}, Springer: New York.
}
\keyword{datasets}
