% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapers.R
\name{dpss.taper}
\alias{dpss.taper}
\title{Calculating Thomson's Spectral Multitapers by Inverse Iteration}
\usage{
dpss.taper(n, k, nw = 4)
}
\arguments{
\item{n}{length of data taper(s)}

\item{k}{number of data tapers; 1, 2, 3, ... (do not use 0!)}

\item{nw}{product of length and half-bandwidth parameter (w)}
}
\value{
\item{v}{matrix of data tapers (cols = tapers)}
\item{eigen}{eigenvalue associated with each data taper, discarded}
}
\description{
This is now a wrapper to the function multitaper::dpss().
}
\references{
Percival, D. B. and A. T. Walden (1993) \emph{Spectral Estimation for
Physical Applications: Multitaper and Conventional Univariate Techniques},
Cambridge University Press.
}
\seealso{
\code{\link{sine.taper}}.
}
\author{
B. Whitcher
}
\keyword{ts}
