% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualtree.R
\name{dualfilt1}
\alias{dualfilt1}
\alias{AntonB}
\title{Kingsbury's Q-filters for the Dual-Tree Complex DWT}
\usage{
dualfilt1()
}
\value{
\item{af}{List (i=1,2) - analysis filters for tree i} \item{sf}{List
(i=1,2) - synthesis filters for tree i} Note: \code{af[[2]]} is the reverse
of \code{af[[1]]}.
}
\description{
Kingsbury's Q-filters for the dual-tree complex DWT.
}
\details{
These cofficients are rounded to 8 decimal places.
}
\references{
Kingsbury, N.G. (2000).  A dual-tree complex wavelet transform
with improved orthogonality and symmetry properties, \emph{Proceedings of
the IEEE Int. Conf. on Image Proc.} (ICIP).
}
\seealso{
\code{\link{dualtree}}
}
\author{
Matlab: S. Cai, K. Li and I. Selesnick; R port: B. Whitcher
}
\keyword{ts}
