% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exchange}
\alias{exchange}
\title{Exchange Rates Between the Deutsche Mark, Japanese Yen and U.S. Dollar}
\format{
A bivariate time series containing 348 observations.
}
\source{
Unknown.
}
\usage{
data(exchange)
}
\description{
Monthly foreign exchange rates for the Deutsche Mark - U.S. Dollar (DEM-USD)
and Japanese Yen - U.S. Dollar (JPY-USD) starting in 1970.
}
\references{
Gencay, R., F. Selcuk and B. Whitcher (2001) \emph{An
Introduction to Wavelets and Other Filtering Methods in Finance and
Economics}, Academic Press.
}
\keyword{datasets}
