% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualtree.R
\name{FSfarras}
\alias{FSfarras}
\alias{farras}
\title{Farras nearly symmetric filters}
\usage{
FSfarras()
}
\value{
\item{af}{List (i=1,2) - analysis filters for tree i} \item{sf}{List
(i=1,2) - synthesis filters for tree i}
}
\description{
Farras nearly symmetric filters for orthogonal 2-channel perfect
reconstruction filter bank and Farras filters organized for the dual-tree
complex DWT.
}
\references{
A. F. Abdelnour and I. W. Selesnick. \dQuote{Nearly symmetric
orthogonal wavelet bases}, Proc. IEEE Int. Conf. Acoust., Speech, Signal
Processing (ICASSP), May 2001.
}
\seealso{
\code{\link{afb}}, \code{\link{dualtree}}, \code{\link{dualfilt1}}.
}
\author{
Matlab: S. Cai, K. Li and I. Selesnick; R port: B. Whitcher
}
\keyword{ts}
