% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heavisine}
\alias{heavisine}
\title{Sine with Jumps at 0.3 and 0.72}
\format{
A vector containing 512 observations.
}
\source{
S+WAVELETS.
}
\usage{
data(heavisine)
}
\description{
\deqn{heavisine(x) = 4\sin(4{\pi}x) - \mathrm{sign}(x-0.3) - }{%
heavisine(x) = 4*sin(4*pi*x) - sign(x-0.3) - sign(0.72-x)}\deqn{
\mathrm{sign}(0.72-x)}{% heavisine(x) = 4*sin(4*pi*x) - sign(x-0.3) -
sign(0.72-x)}
}
\references{
Bruce, A., and H.-Y. Gao (1996) \emph{Applied Wavelet Analysis
with S-PLUS}, Springer: New York.
}
\keyword{datasets}
