% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ibm}
\alias{ibm}
\title{Daily IBM Stock Prices}
\format{
A vector containing 369 observations.
}
\source{
Box, G. E.P. and Jenkins, G. M. (1976) \emph{Time Series Analysis:
Forecasting and Control}, Holden Day, San Francisco, 2nd edition.
}
\usage{
data(ibm)
}
\description{
Daily IBM stock prices spanning May~17, 1961 to November~2, 1962.
}
\keyword{datasets}
