% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kobe}
\alias{kobe}
\title{1995 Kobe Earthquake Data}
\format{
A vector containing 3048 observations.
}
\source{
Data management centre, Washington University.
}
\usage{
data(kobe)
}
\description{
Seismograph (vertical acceleration, nm/sq.sec) of the Kobe earthquake,
recorded at Tasmania University, HobarTRUE, Australia on 16 January 1995
beginning at 20:56:51 (GMTRUE) and continuing for 51 minutes at 1 second
intervals.
}
\keyword{datasets}
