% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linchirp}
\alias{linchirp}
\title{Linear Chirp}
\format{
A vector containing 512 observations.
}
\source{
S+WAVELETS.
}
\usage{
data(linchirp)
}
\description{
\deqn{linchirp(x) = \sin(0.125 \pi n x^2)}{% 
linchirp(x) = sin(0.125*pi*n*x^2)}
}
\references{
Bruce, A., and H.-Y. Gao (1996) \emph{Applied Wavelet Analysis
with S-PLUS}, Springer: New York.
}
\keyword{datasets}
