% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mexm}
\alias{mexm}
\title{Mexican Money Supply}
\format{
A vector containing 516 observations.
}
\source{
Unknown.
}
\usage{
data(mexm)
}
\description{
Percentage changes in monthly Mexican money supply.
}
\references{
Gencay, R., F. Selcuk and B. Whitcher (2001) \emph{An
Introduction to Wavelets and Other Filtering Methods in Finance and
Economics}, Academic Press.
}
\keyword{datasets}
