\name{nile}
\alias{nile}
\title{Nile River Minima}
\description{
  Yearly minimal water levels of the Nile river for the years 622 to
  1281, measured at the Roda gauge near Cairo (Tousson, 1925,
  p. 366-385). The data are listed in chronological sequence by row.

  The original Nile river data supplied by Beran only contained only 500
  observations (622 to 1121).  However, the book claimed to have 660
  observations (622 to 1281).  The remaining observations from the book
  were added, by hand, but the series still only contained 653
  observations (622 to 1264).

  Note, now the data consists of 663 observations (spanning the years
  622-1284) as in original source (Toussoun, 1925).
}
\usage{data(nile)
}
\format{A length 663 vector.
}
\source{
  Toussoun, O. (1925)
  M\'emoire sur l'Histoire du Nil,
  Volume 18 in \emph{M\'emoires a l'Institut d'Egypte},
  pp. 366-404.
}
\references{
  Beran, J. (1994)
  \emph{Statistics for Long-Memory Processes}, Chapman Hall: Englewood, NJ.
}
\keyword{datasets}
