% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{per}
\alias{per}
\title{Periodogram}
\usage{
per(z)
}
\arguments{
\item{z}{time series}
}
\description{
Computation of the periodogram via the Fast Fourier Transform (FFT).
}
\author{
Author: Jan Beran; modified: Martin Maechler, Date: Sep 1995.
}
\keyword{ts}
