% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapers.R
\name{sine.taper}
\alias{sine.taper}
\title{Computing Sinusoidal Data Tapers}
\usage{
sine.taper(n, k)
}
\arguments{
\item{n}{length of data taper(s)}

\item{k}{number of data tapers}
}
\value{
A vector or matrix of data tapers (cols = tapers).
}
\description{
Computes sinusoidal data tapers directly from equations.
}
\details{
See reference.
}
\references{
Riedel, K. S. and A. Sidorenko (1995) Minimum bias multiple
taper spectral estimation, \emph{IEEE Transactions on Signal Processing},
\bold{43}, 188-195.
}
\author{
B. Whitcher
}
\keyword{ts}
