% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tourism}
\alias{tourism}
\title{U.S. Tourism}
\format{
A vector containing 160 observations.
}
\source{
Unknown.
}
\usage{
data(tourism)
}
\description{
Quarterly U.S. tourism figures from 1960:1 to 1999:4.
}
\references{
Gencay, R., F. Selcuk and B. Whitcher (2001) \emph{An
Introduction to Wavelets and Other Filtering Methods in Finance and
Economics}, Academic Press.
}
\keyword{datasets}
