% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unemploy}
\alias{unemploy}
\title{U.S. Unemployment}
\format{
A vector containing 624 observations.
}
\source{
Unknown.
}
\usage{
data(unemploy)
}
\description{
Monthly U.S. unemployment figures from 1948:1 to 1999:12.
}
\references{
Gencay, R., F. Selcuk and B. Whitcher (2001) \emph{An
Introduction to Wavelets and Other Filtering Methods in Finance and
Economics}, Academic Press.
}
\keyword{datasets}
