\name{ar1}
\alias{ar1}
\title{Simulated AR(1) Series}
\description{
  Simulated AR(1) series used in Gencay, Selcuk and Whitcher (2001).
}
\usage{data(ar1)
}
\format{A vector containing 200 observations.
}
%\source{}
\references{
  Gencay, R., F. Selcuk and B. Whitcher (2001)
  \emph{An Introduction to Wavelets and Other Filtering Methods in
    Finance and Economics},
  Academic Press.
}
\keyword{datasets}
