\name{cpi}
\alias{cpi}
\title{U.S. Consumer Price Index}
\description{
  Monthly U.S. consumer price index from 1948:1 to 1999:12.
}
\usage{data(cpi)
}
\format{A vector containing 624 observations.
}
\source{Unknown.
}
\references{
  Gencay, R., F. Selcuk and B. Whitcher (2001)
  \emph{An Introduction to Wavelets and Other Filtering Methods in
    Finance and Economics},
  Academic Press.
}
\keyword{datasets}
