\name{doppler}
\alias{doppler}
\title{Sinusoid with Changing Amplitude and Frequency}
\description{
  \deqn{doppler(x) = \sqrt{x(1 - x)}
    \sin\left(\frac{2.1\pi}{x+0.05}\right)}{%
    doppler(x) = sqrt{x(1-x)} sin[(2.1*pi)/(x+0.05)]}
}
\usage{data(doppler)
}
\format{A vector containing 512 observations.
}
\source{S+WAVELETS.
}
\references{
  Bruce, A., and H.-Y. Gao (1996)
  \emph{Applied Wavelet Analysis with S-PLUS}, Springer: New York.
}
\keyword{datasets}
