\name{ibm}
\alias{ibm}
\title{Daily IBM Stock Prices}
\description{
  Daily IBM stock prices spanning May~17, 1961 to November~2, 1962.
}
\usage{data(ibm)
}
\format{A vector containing 369 observations.
}
\source{
  Box, G. E.~P. and Jenkins, G.~M. (1976)
  \emph{Time Series Analysis: Forecasting and Control},
  Holden Day, San Francisco, 2 edition.
}
\references{
  \url{http://www-personal.buseco.monash.edu.au/~hyndman/TSDL/}
}
\keyword{datasets}
