\name{japan}
\alias{japan}
\title{Japanese Gross National Product}
\description{
  Quarterly Japanese gross national product from 1955:1 to 1996:4.
}
\usage{data(japan)
}
\format{A vector containing 169 observations.
}
\source{Unknown.
}
\references{
  Gencay, R., F. Selcuk and B. Whitcher (2001)
  \emph{An Introduction to Wavelets and Other Filtering Methods in
    Finance and Economics},
  Academic Press.

  Hecq, A. (1998)
  Does seasonal adjustment induce common cycles?,
  \emph{Empirical Economics}, \bold{59}, 289-297.
}
\keyword{datasets}
