\name{linchirp}
\alias{linchirp}
\title{Linear Chirp}
\description{
  \deqn{linchirp(x) = \sin(0.125 \pi n x^2)}{%
    linchirp(x) = sin(0.125*pi*n*x^2)}
}
\usage{data(linchirp)
}
\format{A vector containing 512 observations.
}
\source{S+WAVELETS.
}
\references{
  Bruce, A., and H.-Y. Gao (1996)
  \emph{Applied Wavelet Analysis with S-PLUS}, Springer: New York.
}
\keyword{datasets}
