\name{mexm}
\alias{mexm}
\title{Mexican Money Supply}
\description{
  Percentage changes in monthly Mexican money supply.
}
\usage{data(mexm)
}
\format{A vector containing 516 observations.
}
\source{Unknown.
}
\references{
  Gencay, R., F. Selcuk and B. Whitcher (2001)
  \emph{An Introduction to Wavelets and Other Filtering Methods in
    Finance and Economics},
  Academic Press.
}
\keyword{datasets}
