\name{per}
\alias{per}
\title{Periodogram}
\description{
  Computation of the periodogram via the Fast Fourier Transform (FFT).
}
\usage{per(z)
}
\arguments{
  \item{z}{time series}
}
%\value{}
%\details{}
%\references{}
%\seealso{}
%\examples{}
\author{Author: Jan Beran;  modified: Martin Maechler, Date: Sep 1995.}
\keyword{ts}
