\name{Selesnick}
\alias{cshift}
\alias{cshift2D}
\alias{pm}
\title{Miscellaneous Functions for Dual-tree Wavelet Software}
\description{
  
}
\usage{
cshift(x, m)
cshift2D(x, m)
pm(a, b)
}
\arguments{
  \item{x}{$N$-point vector}
  \item{m}{amount of shift}
  \item{a,b}{}
}
\value{
  \item{y}{vector $x$ will be shifed by $m$ samples to the left or
    matrix $x$ will be shifed by $m$ samples down.}
}
\details{
}
\references{
  WAVELET SOFTWARE AT POLYTECHNIC UNIVERSITY, BROOKLYN, NY
  http://taco.poly.edu/WaveletSoftware/
}
%\seealso{}
%\examples{}
\author{Matlab: S. Cai, K. Li and I. Selesnick; R port: B. Whitcher}
\keyword{ts}
