\name{tourism}
\alias{tourism}
\title{U.S. Tourism}
\description{
  Quarterly U.S. tourism figures from 1960:1 to 1999:4.
}
\usage{data(tourism)
}
\format{A vector containing 160 observations.
}
\source{Unknown.
}
\references{
  Gencay, R., F. Selcuk and B. Whitcher (2001)
  \emph{An Introduction to Wavelets and Other Filtering Methods in
    Finance and Economics},
  Academic Press.
}
\keyword{datasets}
