\name{unemploy}
\alias{unemploy}
\title{U.S. Unemployment}
\description{
  Monthly U.S. unemployment figures from 1948:1 to 1999:12.
}
\usage{data(unemploy)
}
\format{A vector containing 624 observations.
}
\source{Unknown.
}
\references{
  Gencay, R., F. Selcuk and B. Whitcher (2001)
  \emph{An Introduction to Wavelets and Other Filtering Methods in
    Finance and Economics},
  Academic Press.
}
\keyword{datasets}
